requirements.txt
# CrisisWatch - Crisis Misinformation Detection Agent
# Core dependencies

# LLM Framework
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# LLM Providers
google-generativeai>=0.8.0
openai>=1.50.0  # Used for Grok (xAI uses OpenAI-compatible API)

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.10.0

# Data validation
pydantic>=2.9.0
pydantic-settings>=2.5.0

# NLP
spacy>=3.7.0

# Environment & Config
python-dotenv>=1.0.0

# CLI & UI
rich>=13.9.0
typer>=0.12.0
streamlit>=1.40.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# API integrations
tavily-python>=0.5.0

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
